+++ tests/testthat/test-hhcf.R
@@ -1,0 +1,8 @@
+test_that("Height height correlation function: check roughness", {
+  filename = AFM.getSampleImages(type='tiff')
+  a = AFM.import(filename)
+  a = AFM.flatten(a)
+  h1 = AFM.hhcf(a, numIterations = 4e4, degRes=1000, dataOnly = TRUE)
+
+  expect_equal(mean(h1$g) / (2*AFM.math.params(a)$Rq^2), 1, tolerance = 2e-2)
+})